+++ tsup.config.ts
@@ -3,8 +3,6 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     index: "src/index.ts",
-    "presentation/index": "src/presentation/index.ts",
-    "problem-details/index": "src/problem-details/index.ts",
     "public-error/index": "src/public-error/index.ts",
   },
   format: ["esm", "cjs"],